[BOLT-TESTS] Add libmysqlclient-entry-cold.test
Summary:
Previous assumption is that a cold fragment can only be reached through
another main fragment, and these cold fragments are already handled.
However, this assumption is too strict, e.g., it breaks libmysqlclient.

This test includes (a) libmysqlclient, and (b) a manual test to check
that cold fragments are allowed to be function entry and program entry.